+++ packages/swapper/src/swappers/zrx/ZrxSwapper.test.ts
@@ -7,11 +7,13 @@ import { ZrxError } from '../..'
 import { ZrxSwapper } from '..'
 import { buildQuoteTx } from '../zrx/buildQuoteTx/buildQuoteTx'
 import { executeQuote } from '../zrx/executeQuote/executeQuote'
+import { approvalNeeded } from './approvalNeeded/approvalNeeded'
+import { approveInfinite } from './approveInfinite/approveInfinite'
 import { getMinMax } from './getMinMax/getMinMax'
 import { getZrxQuote } from './getQuote/getQuote'
-import { DEFAULT_SLIPPAGE } from './utils/constants'
 import { getUsdRate } from './utils/helpers/helpers'
 import { BTC, FOX, WETH } from './utils/test-data/assets'
+import { setupQuote } from './utils/test-data/setupSwapQuote'
 
 jest.mock('./utils/helpers/helpers')
 jest.mock('../zrx/executeQuote/executeQuote', () => ({
@@ -30,19 +32,13 @@ jest.mock('./getMinMax/getMinMax', () => ({
   getMinMax: jest.fn()
 }))
 
-const setupQuote = () => {
-  const sellAmount = '1000000000000000000'
-  const sellAsset = FOX
-  const buyAsset = WETH
+jest.mock('./approvalNeeded/approvalNeeded', () => ({
+  approvalNeeded: jest.fn()
+}))
 
-  const quoteInput = {
-    sellAsset,
-    buyAsset,
-    sellAmount,
-    slippage: DEFAULT_SLIPPAGE
-  }
-  return { quoteInput, buyAsset, sellAsset }
-}
+jest.mock('./approveInfinite/approveInfinite', () => ({
+  approveInfinite: jest.fn()
+}))
 
 describe('ZrxSwapper', () => {
   const input = <GetQuoteInput>{}
@@ -113,5 +109,20 @@ describe('ZrxSwapper', () => {
     const { quoteInput } = setupQuote()
     await swapper.getMinMax(quoteInput)
     expect(getMinMax).toHaveBeenCalled()
+  })
+
+  it('calls approvalNeeded on swapper.approvalNeeded', async () => {
+    const swapper = new ZrxSwapper(zrxSwapperDeps)
+    const { quoteInput } = setupQuote()
+    const args = { quote: quoteInput, wallet }
+    await swapper.approvalNeeded(args)
+    expect(approvalNeeded).toHaveBeenCalled()
+  })
+
+  it('calls approveInfinite on swapper.approveInfinite', async () => {
+    const swapper = new ZrxSwapper(zrxSwapperDeps)
+    const args = { quote, wallet }
+    await swapper.approveInfinite(args)
+    expect(approveInfinite).toHaveBeenCalled()
   })
 })